requirements.txt
# Core dependencies
pandas>=2.0.0
matplotlib>=3.10.7
pyyaml>=6.0

# Web scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# Optional: for better performance
lxml>=4.9.0
beautifulsoup4>=4.12.0

# Development dependencies (optional, install with: pip install -r requirements-dev.txt)
# pytest>=7.3.0
# pytest-cov>=4.1.0
# black>=23.3.0
# flake8>=6.0.0